refactor(ts): move to TypeScript release of yargs-parser
