+++ lib/validation.ts
@@ -4,7 +4,7 @@ import { levenshtein as distance } from './levenshtein'
 import { objFilter } from './obj-filter'
 import { UsageInstance } from './usage'
 import { YargsInstance, Arguments } from './yargs'
-import { DetailedArguments } from 'yargs-parser'
+import { DetailedArguments } from 'yargs-parser/build/lib/yargs-parser-types'
 import { Y18N } from 'y18n'
 const specialKeys = ['$0', '--', '_']
 
@@ -131,8 +131,8 @@ export function validation (yargs: YargsInstance, usage: UsageInstance, y18n: Y1
 
     if ((currentContext.commands.length > 0) || (commandKeys.length > 0) || isDefaultCommand) {
       argv._.slice(currentContext.commands.length).forEach((key) => {
-        if (commandKeys.indexOf(key) === -1) {
-          unknown.push(key)
+        if (commandKeys.indexOf('' + key) === -1) {
+          unknown.push('' + key)
         }
       })
     }
@@ -154,8 +154,8 @@ export function validation (yargs: YargsInstance, usage: UsageInstance, y18n: Y1
 
     if ((currentContext.commands.length > 0) || (commandKeys.length > 0)) {
       argv._.slice(currentContext.commands.length).forEach((key) => {
-        if (commandKeys.indexOf(key) === -1) {
-          unknown.push(key)
+        if (commandKeys.indexOf('' + key) === -1) {
+          unknown.push('' + key)
         }
       })
     }
